backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0

# OpenAI API
openai>=1.3.0

# Database (optional, for future persistence)
# sqlalchemy>=2.0.0
# aiosqlite>=0.19.0

# Image and export
python-pptx>=0.6.21
Pillow>=10.0.0

# Browser automation for screenshots (uses system Chrome, inherits system fonts)
selenium>=4.15.0
webdriver-manager>=4.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0